[Runners] Add events to notify that a test run started and completed. (#46)
Add two events that notify when the runner started and when it
completed. The completed event will provide information of the results
of the run.

Later we can look at adding event for each of the tests starting and
completing, which is possible, but we are not going to add events that
we have not been asked to add. KISS

fixes: https://github.com/dotnet/xharness/issues/36