Increase the timeout for listing simulators (#298)
Through investigation, I found out that mlaunch takes 2-4 minutes to list devices on a cold start. After that, some cache is in effect and listing of devices takes 1-30 seconds.
If we do not complete the run, next attempt requires the same amount of time. So if the timeout wouldn't be sufficient, retry won't help.

Increasing this timeout will prevent occasional problems. I will also be deleting all created devices from our infrastructure to speed things up but it doesn't solve the underlying issue completely and that is still up for an investigation.