[wasm] Log browser level errors, instead of interrupting the (#311)
.. tests by throwing an exception.

Errors like:

`http://bla/ - Failed to load resource: net::ERR_NAME_NOT_RESOLVED)`

Note: I wasn't able to capture these in js itself, to pipe through the
console websocket. So, we get these from Selenium, and that can be out
of sync with the messages from the console websocket.

This was seen in tests for `System.XmlSchemaSet.Tests`. These errors
would get logged, but since we were throwing exceptions for them, the
result of all the tests would marked as `GENERAL_FAILURE`.

Note: In this case the errors are expected, but I'm not sure how to
differentiate these from other browser level errors that mean that the
whole thing has failed.