Remove the `ios package` command (#342)
The package command was never finished, it is not maintained, it is not tested and packages are built by other tools in `dotnet/runtime`. If we ever choose to do that, we will re-use those tools or integrate them in XHarness.

It will be better if this code stays in Xamarin (where it changed already in the meantime) and Xamarin guys have full control over it. I imagine it will be way easier for them to make changes there.