Handle app run launch failure in `ios run` (#421)
When we fail to run the application in `ios test`, we return with 83 which is a signal for Arcade SDK to retry/reboot the work item.

For `ios run` this was broken and this was not detected at all. Further, it was still being compared to the value in `--expected-exit-code` ending up with a failure.