Separate install and run operations for iOS apps (#538)
In order to introduce the new `install` / `run-test` / `uninstall` commands (#493), we need to separate installation from running the app which is currently happening via 1 mlaunch command for Simulators.

We unify the approach for Simulators and devices and do the install separately. This will also help with diagnosis of issues when simulators freeze and we don't know at which stage.

The code was refactored out of the `Command` classes into the `Orchestrator` classes so that it can be tested later too.
We can also now test passing args to the `Commands` and how they get parsed.