Add `--signal-app-end` to detect test run end for iOS/tvOS 14+ devices (#643)
Signals the test runner to log a given tag at the end of the test run and ends the test run. Useful for iOS/tvOS 14+ where mlaunch cannot detect the app quit.

This functionality should be eventually superseded by a full protocol between the app and XHarness but this way keeps it backwards compatible and also won't break Xamarin who is sharing some of the code.