ThreadlessXunitTestRunner: Escape newlines in test result lines (#817)
`ThreadlessXunitTestRunner` writes lines like:

    `[PASS] foo.testname(testargs)`

to the console, which gets parsed by `WasmTestMessageProcessor`. But if
the list of arguments contains a new line(eg. in data for
`System.Text.RegularExpressions` tests), then the output line gets written
as multiple lines to the console. And that results in the output looking
like:

```
[02:41:04] info: abc, beginning: 0, length: 5, expectedSuccess: True, expectedValue: "a\nabc")
[02:41:04] info: abc, beginning: 0, length: 5, expectedSuccess: True, expectedValue: "a\nabc")
...
```

Each of those lines being the last part of the `[PASS] foo.testname(someargs, "\nabc, ..")`.

Instead, we escape the new line chars.

Partially fixes https://github.com/dotnet/xharness/issues/816